move checkbox-list
