+++ src/packages/core/property-editor/uis/index.ts
@@ -1,2 +1,2 @@
 // TODO: Currently we export input-checkbox-list because package-builder uses it, but this should be changed over time, so package builder uses the property editor UI and then this export can be removed.
-export * from './checkbox-list/index.js';
+export * from '../../../property-editors/checkbox-list/index.js';
