+++ src/packages/property-editors/checkbox-list/manifests.ts
