+++ src/packages/property-editors/checkbox-list/property-editor-ui-checkbox-list.element.ts
